Added view for updating Stock Levels

+++ db/migrate/20131112100450_create_items.rb
@@ -2,7 +2,7 @@ class CreateItems < ActiveRecord::Migration
   def change
     create_table :items do |t|
       t.string :name
-      t.integer :price
+      t.double :price
       t.string :details
       t.integer :quantity
       t.string :supplier
